+++ src/CustomPropertyTypes/utils.ts
@@ -10,6 +10,20 @@ import { customPropertyTypePrefix } from "~/lib/constants";
 import { MetadataTypeManager } from "obsidian-typings";
 import { NonNullishObject } from "~/lib/utils";
 
+/**
+ * Array sub-properties are in the format `<parent>.<index>`. This function replaces the index with a `#` to properly get/set settings for the sub-property
+ */
+export const getTrueProperty = (property: string) => {
+	return property
+		.split(".")
+		.map((part) => {
+			const n = Number(part);
+			if (Number.isNaN(n)) return part;
+			return "#";
+		})
+		.join(".");
+};
+
 export const getPropertySettings = ({
 	plugin,
 	property,
@@ -17,7 +31,7 @@ export const getPropertySettings = ({
 	plugin: BetterProperties;
 	property: string;
 }): NonNullishObject<PropertySettings> => {
-	const lower = property.toLowerCase();
+	const lower = getTrueProperty(property).toLowerCase();
 	if (!plugin.settings.propertySettings) {
 		plugin.settings.propertySettings = {};
 	}
@@ -49,7 +63,7 @@ export const deletePropertySettings = ({
 	plugin: BetterProperties;
 	property: string;
 }) => {
-	const lower = property.toLowerCase();
+	const lower = getTrueProperty(property).toLowerCase();
 	if (!plugin.settings.propertySettings) {
 		plugin.settings.propertySettings = {};
 	}
@@ -66,7 +80,7 @@ export const setPropertySettings = ({
 	property: string;
 	settings: PropertySettings;
 }): void => {
-	const lower = property.toLowerCase();
+	const lower = getTrueProperty(property).toLowerCase();
 	if (!plugin.settings.propertySettings) {
 		plugin.settings.propertySettings = {};
 	}