+++ src/Plugin/plugin.ts
@@ -1,4 +1,10 @@
-import { Modal, ButtonComponent, Plugin, MarkdownView } from "obsidian";
+import {
+	Modal,
+	ButtonComponent,
+	Plugin,
+	MarkdownView,
+	MetadataTypeManager,
+} from "obsidian";
 import {
 	BetterPropertiesSettings,
 	betterPropertiesSettingsSchema,
@@ -23,6 +29,31 @@ import * as v from "valibot";
 import { openRenameModal } from "~/MetadataEditor/propertyEditorMenu/rename";
 import { registerBpJsCodeProcessors } from "~/bpjs";
 import { BpJsApi, setupBpJsListeners } from "~/bpjs/api";
+import { around, dedupe } from "monkey-around";
+import { monkeyAroundKey } from "~/lib/constants";
+import { getTrueProperty } from "~/CustomPropertyTypes/utils";
+
+const patchMetadataTypeManager = (plugin: BetterProperties): void => {
+	const removePatch = around(plugin.app.metadataTypeManager, {
+		getAssignedWidget(old) {
+			return dedupe(monkeyAroundKey, old, function (property) {
+				// @ts-expect-error
+				const that: MetadataTypeManager = this;
+				return old.call(that, getTrueProperty(property));
+			});
+		},
+		setType(old) {
+			return dedupe(monkeyAroundKey, old, function (property, type) {
+				// @ts-expect-error
+				const that: MetadataTypeManager = this;
+				return old.call(that, getTrueProperty(property), type);
+			});
+		},
+	});
+
+	plugin.register(removePatch);
+};
+
 export class BetterProperties extends Plugin {
 	settings: BetterPropertiesSettings = getDefaultSettings();
 	disabledTypeWidgets: Record<string, PropertyWidget> = {};
@@ -33,6 +64,7 @@ export class BetterProperties extends Plugin {
 	async onload(): Promise<void> {
 		await this.loadSettings();
 		this.addSettingTab(new BetterPropertiesSettingsTab(this));
+		patchMetadataTypeManager(this);
 		registerCustomPropertyTypeWidgets(this);
 		wrapAllPropertyTypeWidgets(this);
 		sortAndFilterRegisteredTypeWidgets(this);