+++ changelog.md
@@ -12,6 +12,7 @@ release. Some specific details may be omitted for brevity.
 ### Fixed
 
 - Hidden properties visiblity has been fixed when the "Show hidden" button is toggled on.
+- ColorTextComponent's hidden color input is now correctly positioned within it's parent.
 
 ### Other