+++ changelog.md
@@ -3,6 +3,20 @@
 Each release should include an entry detailing the changes made since the last
 release. Some specific details may be omitted for brevity.
 
+## 0.1.11
+
+### New
+
+- Property value collapse indicators now display only when hovering over the type icon. [#30](https://github.com/unxok/obsidian-better-properties/issues/30).
+
+### Fixed
+
+- ...
+
+### Other
+
+- ...
+
 ## 0.1.10
 
 This should be the last release before the beta annoucement🎉